requirements.txt
atomicwrites==1.4.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.6
Django==2.2.16
django-debug-toolbar==3.2.4
Faker==12.0.1
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
mixer==7.1.2
packaging==23.0
Pillow==8.3.1
pluggy==0.13.1
py==1.11.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
